Adds RESTful end point for /users.
This update modifies the way the /users endpoint data is stored and
fetched. Rather than pushing all the data into a single array inside a
key for each user, a new key is created for each favorite. This makes it
easier to send a delete command to remove a single favorite. It also
allows us to filter the list of items returned from a GET request, if
parameters like page number / page limit are passed in.